+++ Solutions/CrowdStrike Falcon Endpoint Protection/Data Connectors/CrowdstrikeReplicatorCLv2/CrowdstrikeFalconAPISentinelConn_Premium/QueueTriggerCS/readme.md
@@ -1,0 +1,11 @@
+# QueueTrigger - Python
+
+The `QueueTrigger` makes it incredibly easy to react to new Queues inside of Azure Queue Storage. This sample demonstrates a simple use case of processing data from a given Queue.
+
+## How it works
+
+For a `QueueTrigger` to work, you provide a path which dictates where the queue messages are located inside your container.
+
+## Learn more
+
+<TODO> Documentation
