+++ Solutions/CrowdStrike Falcon Endpoint Protection/Data Connectors/CrowdstrikeReplicatorCLv2/CrowdstrikeFalconAPISentinelConn_Premium/requirements.txt
@@ -1,0 +1,5 @@
+azure-storage-queue==12.1.5
+azure-core==1.27.1
+azure-functions==1.6.0
+gzip_stream==1.2.0
+aiobotocore==2.5.1
