--- Solutions/CrowdStrike Falcon Endpoint Protection/Data Connectors/CrowdstrikeReplicatorCLv2/Replicator/function.json
@@ -1,12 +1,0 @@
-{
-    "scriptFile": "main_aws_queue.py",
-    "bindings": [
-      {
-        "name": "mytimer",
-        "type": "timerTrigger",
-        "direction": "in",
-        "schedule": "%Schedule%",
-        "runOnStartup": false
-      }
-    ]
-  }
