--- Solutions/CrowdStrike Falcon Endpoint Protection/Data Connectors/CrowdstrikeReplicatorCLv2/host.json
@@ -1,26 +1,0 @@
-{
-  "version": "2.0",
-  "functionTimeout": "00:10:00",
-  "logging": {
-    "applicationInsights": {
-      "samplingSettings": {
-        "isEnabled": true,
-        "excludedTypes": "Request"
-      }
-    }
-  },
-  "extensionBundle": {
-    "id": "Microsoft.Azure.Functions.ExtensionBundle",
-    "version": "[3.*, 4.0.0)"
-  },
-  "extensions": {
-    "queues": {
-      "maxPollingInterval": "00:01:00",
-      "visibilityTimeout": "00:00:30",
-      "batchSize": 8,
-      "maxDequeueCount": 5,
-      "newBatchThreshold": 4,
-      "messageEncoding": "base64"
-    }
-  }
-}
