--- Solutions/CrowdStrike Falcon Endpoint Protection/Data Connectors/CrowdstrikeReplicatorCLv2/proxies.json
@@ -1,4 +1,0 @@
-{
-  "$schema": "http://json.schemastore.org/proxies",
-  "proxies": {}
-}
